# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Comptes",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes",
                "description": "Récupère la liste des comptes avec pagination et filtres",
                "operationId": "getComptes",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de compte (epargne, cheque)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Statut du compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par titulaire ou numéro de compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "dateCreation",
                            "enum": [
                                "dateCreation",
                                "solde",
                                "titulaire"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Compte"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/Meta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "succes": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Compte": {
                "title": "Compte",
                "description": "Objet représentant un compte bancaire",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numeroCompte": {
                        "type": "string",
                        "example": "ACC-ABC123DEF"
                    },
                    "titulaire": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ],
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1500.5
                    },
                    "devise": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "dateCreation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-23T10:00:00Z"
                    },
                    "statut": {
                        "type": "string",
                        "example": "actif"
                    },
                    "motifBlocage": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2023-10-23T10:00:00Z"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Meta": {
                "title": "Meta",
                "description": "Informations de pagination et liens",
                "properties": {
                    "pagination": {
                        "properties": {
                            "currentPage": {
                                "type": "integer",
                                "example": 1
                            },
                            "totalPages": {
                                "type": "integer",
                                "example": 5
                            },
                            "totalItems": {
                                "type": "integer",
                                "example": 50
                            },
                            "itemsPerPage": {
                                "type": "integer",
                                "example": 10
                            },
                            "hasNext": {
                                "type": "boolean",
                                "example": true
                            },
                            "hasPrevious": {
                                "type": "boolean",
                                "example": false
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=1"
                            },
                            "next": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=2"
                            },
                            "first": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=1"
                            },
                            "last": {
                                "type": "string",
                                "example": "http://localhost:8000/api/v1/comptes?page=5"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Comptes"
        }
    ]
}